separate conformance utilities
